Make a complex128, not 64

+++ spectral_connectivity/connectivity.py
@@ -353,7 +353,7 @@ class Connectivity:
         # U - rotation
         unnormalized_global_coherence = np.zeros(
             (n_time_windows, n_fft_samples, n_signals, max_rank),
-            dtype=np.complex64)
+            dtype=np.complex128)
 
         for time_ind in range(n_time_windows):
             for freq_ind in range(n_fft_samples):
